tests: model: Fix thumbs_up emoji code in parametrized index argument.
This fixes thumbs_up emoji code in
test__handle_reaction_event_remove_reaction's index argument to be as
that of response and amends the assert statement to pass the test.

The amendment in the assert statement shows how any reaction (without
matching the user ids) is being removed in the code (see #606).